+++ adminsec/tests/test_views.py
@@ -30,6 +30,7 @@ from usersec.tests.factories import (
     HpcUserCreateRequestFactory,
     HpcProjectCreateRequestFactory,
     HpcUserFactory,
+    HpcGroupChangeRequestFactory,
 )
 from usersec.tests.test_views import TestViewBase
 
@@ -326,6 +327,267 @@ class TestHpcGroupCreateRequestDenyView(TestViewBase):
                 reverse(
                     "adminsec:hpcgroupcreaterequest-deny",
                     kwargs={"hpcgroupcreaterequest": self.obj.uuid},
+                ),
+                data={"comment": "Denied!"},
+            )
+
+            self.assertRedirects(
+                response,
+                reverse("adminsec:overview"),
+            )
+
+            messages = list(get_messages(response.wsgi_request))
+            self.assertEqual(len(messages), 1)
+            self.assertEqual(str(messages[0]), "Request successfully denied.")
+
+            self.assertEqual(self.obj.status, REQUEST_STATUS_ACTIVE)
+            self.obj.refresh_from_db()
+            self.assertEqual(self.obj.comment, "Denied!")
+            self.assertEqual(self.obj.status, REQUEST_STATUS_DENIED)
+
+
+class TestHpcGroupChangeRequestDetailView(TestViewBase):
+    """Tests for HpcGroupChangeRequestDetailView."""
+
+    def test_get(self):
+        request = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_ACTIVE
+        )
+
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-detail",
+                    kwargs={"hpcgroupchangerequest": request.uuid},
+                )
+            )
+
+            self.assertEqual(response.status_code, 200)
+            self.assertFalse(response.context["is_decided"])
+            self.assertFalse(response.context["is_denied"])
+            self.assertFalse(response.context["is_retracted"])
+            self.assertFalse(response.context["is_approved"])
+            self.assertTrue(response.context["is_active"])
+            self.assertFalse(response.context["is_revision"])
+            self.assertFalse(response.context["is_revised"])
+            self.assertTrue(response.context["admin"])
+
+    def test_get_retracted(self):
+        request = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_RETRACTED
+        )
+
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-detail",
+                    kwargs={"hpcgroupchangerequest": request.uuid},
+                )
+            )
+
+            self.assertEqual(response.status_code, 200)
+            self.assertFalse(response.context["is_decided"])
+            self.assertFalse(response.context["is_denied"])
+            self.assertTrue(response.context["is_retracted"])
+            self.assertFalse(response.context["is_approved"])
+            self.assertFalse(response.context["is_active"])
+            self.assertFalse(response.context["is_revision"])
+            self.assertFalse(response.context["is_revised"])
+
+    def test_get_denied(self):
+        request = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_DENIED
+        )
+
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-detail",
+                    kwargs={"hpcgroupchangerequest": request.uuid},
+                )
+            )
+
+            self.assertEqual(response.status_code, 200)
+            self.assertTrue(response.context["is_decided"])
+            self.assertTrue(response.context["is_denied"])
+            self.assertFalse(response.context["is_retracted"])
+            self.assertFalse(response.context["is_approved"])
+            self.assertFalse(response.context["is_active"])
+            self.assertFalse(response.context["is_revision"])
+            self.assertFalse(response.context["is_revised"])
+
+    def test_get_approved(self):
+        request = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_APPROVED
+        )
+
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-detail",
+                    kwargs={"hpcgroupchangerequest": request.uuid},
+                )
+            )
+
+            self.assertEqual(response.status_code, 200)
+            self.assertTrue(response.context["is_decided"])
+            self.assertFalse(response.context["is_denied"])
+            self.assertFalse(response.context["is_retracted"])
+            self.assertTrue(response.context["is_approved"])
+            self.assertFalse(response.context["is_active"])
+            self.assertFalse(response.context["is_revision"])
+            self.assertFalse(response.context["is_revised"])
+
+
+class TestHpcGroupChangeRequestRevisionView(TestViewBase):
+    """Tests for HpcGroupChangeRequestRevisionView."""
+
+    def setUp(self):
+        super().setUp()
+        self.obj = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_ACTIVE
+        )
+
+    def test_get(self):
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-revision",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
+                )
+            )
+
+            self.assertEqual(response.status_code, 200)
+            self.assertNotIn("resources_requested", response.context["form"])
+            self.assertNotIn("description", response.context["form"])
+            self.assertNotIn("expiration", response.context["form"])
+            self.assertEqual(response.context["form"]["comment"].value(), "")
+            self.assertTrue(response.context["update"])
+
+    def test_post(self):
+        update = {
+            "comment": "I made a comment!",
+            "resources_requested": json.dumps(self.obj.resources_requested),
+            "description": self.obj.description,
+            "expiration": self.obj.expiration,
+        }
+
+        with self.login(self.user_hpcadmin):
+            response = self.client.post(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-revision",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
+                ),
+                update,
+            )
+            self.assertRedirects(
+                response,
+                reverse("adminsec:overview"),
+            )
+
+            self.obj.refresh_from_db()
+
+            self.assertEqual(self.obj.comment, update["comment"])
+            self.assertEqual(
+                self.obj.resources_requested,
+                json.loads(update["resources_requested"]),
+            )
+            self.assertEqual(self.obj.description, update["description"])
+            self.assertEqual(self.obj.editor, self.user_hpcadmin)
+            self.assertEqual(self.obj.requester, self.user_owner)
+
+            messages = list(get_messages(response.wsgi_request))
+            self.assertEqual(len(messages), 1)
+            self.assertEqual(str(messages[0]), "Revision requested.")
+
+
+class TestHpcGroupChangeRequestApproveView(TestViewBase):
+    """Tests for HpcGroupChangeRequestApproveView."""
+
+    def setUp(self):
+        super().setUp()
+        self.obj = HpcGroupChangeRequestFactory(
+            requester=self.user_owner,
+            group=self.hpc_group,
+            status=REQUEST_STATUS_ACTIVE,
+            delegate=self.hpc_member,
+        )
+
+    def test_get(self):
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-approve",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
+                )
+            )
+            self.assertEqual(response.status_code, 200)
+            self.assertIsNotNone(response.context["form"])
+
+    @override_settings(
+        INSTITUTE_EMAIL_DOMAINS="charite.de",
+        INSTITUTE_USERNAME_SUFFIX="c",
+    )
+    def test_post(self):
+        with self.login(self.user_hpcadmin):
+            response = self.client.post(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-approve",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
+                ),
+            )
+
+            self.assertRedirects(
+                response,
+                reverse("adminsec:overview"),
+            )
+
+            messages = list(get_messages(response.wsgi_request))
+            self.assertEqual(len(messages), 1)
+            self.assertEqual(str(messages[0]), "Request approved and group updated.")
+
+            self.assertEqual(self.obj.status, REQUEST_STATUS_ACTIVE)
+            self.obj.refresh_from_db()
+            self.assertEqual(self.obj.status, REQUEST_STATUS_APPROVED)
+
+            hpcgroups = HpcGroup.objects.all()
+
+            self.assertEqual(hpcgroups.count(), 1)
+
+            hpcgroup = hpcgroups.last()
+
+            self.assertEqual(hpcgroup.delegate, self.hpc_member)
+            self.assertEqual(hpcgroup.description, self.obj.description)
+            self.assertEqual(hpcgroup.resources_requested, self.obj.resources_requested)
+            self.assertEqual(len(mail.outbox), 0)
+
+
+class TestHpcGroupChangeRequestDenyView(TestViewBase):
+    """Tests for HpcGroupChangeRequestDenyView."""
+
+    def setUp(self):
+        super().setUp()
+        self.obj = HpcGroupChangeRequestFactory(
+            requester=self.user_owner, group=self.hpc_group, status=REQUEST_STATUS_ACTIVE
+        )
+
+    def test_get(self):
+        with self.login(self.user_hpcadmin):
+            response = self.client.get(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-deny",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
+                )
+            )
+            self.assertEqual(response.status_code, 200)
+            self.assertIsNotNone(response.context["form"])
+
+    def test_post(self):
+        with self.login(self.user_hpcadmin):
+            response = self.client.post(
+                reverse(
+                    "adminsec:hpcgroupchangerequest-deny",
+                    kwargs={"hpcgroupchangerequest": self.obj.uuid},
                 ),
                 data={"comment": "Denied!"},
             )